requirements.txt
asgiref==3.8.1
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.8
defusedxml==0.8.0rc2
Django==5.0.6
django-compat==1.0.15
django-filter==24.2
django-oauth-toolkit==2.4.0
django-rest-framework-social-oauth2==1.2.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-social-oauth2==2.2.0
drf-yasg==1.21.7
idna==3.7
inflection==0.5.1
jwcrypto==1.5.6
Markdown==3.6
oauthlib==3.2.2
packaging==24.1
pillow==10.3.0
pycparser==2.22
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
requests-oauthlib==2.0.0
six==1.16.0
social-auth-app-django==5.4.1
social-auth-core==4.5.4
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
